+++ module/item/item.js
@@ -912,7 +912,7 @@ export class HeroSystem6eItem extends Item {
             return 0
 
 
-        if (this.system.XMLID === "SWIMMING") {
+        if (this.system.XMLID === "TRANSPORT_FAMILIARITY") {
             console.log(this)
         }
 
@@ -1007,27 +1007,44 @@ export class HeroSystem6eItem extends Item {
                 adderCost += Math.ceil(adderBaseCost * adderLevels)  // ceil is for ENTANGLE +5 PD
             }
 
+            let subAdderCost = 0
+
             for (let adder2 of (adder.ADDER || [])) {
                 let adderBaseCost = adder2.baseCost
 
                 if (adder2.SELECTED != false) {
                     let adderLevels = Math.max(1, parseInt(adder2.LEVELS))
-                    adderCost += Math.ceil(adderBaseCost * adderLevels)
+                    subAdderCost += Math.ceil(adderBaseCost * adderLevels)
                 }
             }
+
+            // TRANSPORT_FAMILIARITY checking more than 2 animals costs same as entire category
+            if (!adder.SELECTED && subAdderCost > adderBaseCost)
+            {
+                subAdderCost = adderBaseCost
+            }
+
+            // Riding discount
+            if (this.system.XMLID === "TRANSPORT_FAMILIARITY" && this.actor && subAdderCost > 0) {
+                if (adder.XMLID === "RIDINGANIMALS" && this.actor.items.find(o=> o.system.XMLID === "RIDING")) {
+                    subAdderCost -= 1
+                }
+            }
+
+            adderCost += subAdderCost
 
         }
 
 
         // Categorized skills cost 2 per catory and +1 per each subcategory.
         // If no catagories selected then assume 3 pts
-        if (configPowerInfo?.categorized && adderCost >= 4) {
-            if (adderCost == 0) {
-                adderCost = 3
-            } else {
-                adderCost = Math.floor(adderCost / 2) + 1
-            }
-        }
+        // if (configPowerInfo?.categorized && adderCost >= 4) {
+        //     if (adderCost == 0) {
+        //         adderCost = 3
+        //     } else {
+        //         adderCost = Math.floor(adderCost / 2) + 1
+        //     }
+        // }
 
 
 
@@ -1077,9 +1094,9 @@ export class HeroSystem6eItem extends Item {
         let advantagesDC = 0;
         let minAdvantage = 0;
 
-        if (this.system.XMLID === "SWIMMING") {
-            console.log(this)
-        }
+        // if (this.system.XMLID === "SWIMMING") {
+        //     console.log(this)
+        // }
 
         for (let modifier of (system.MODIFIER || []).filter(o =>
             (system.XMLID != "NAKEDMODIFIER" || o.PRIVATE)
@@ -1208,9 +1225,9 @@ export class HeroSystem6eItem extends Item {
             limitations += _myLimitation
         }
 
-        if (this.system.XMLID === "SWIMMING") {
-            console.log(this)
-        }
+        // if (this.system.XMLID === "SWIMMING") {
+        //     console.log(this)
+        // }
 
         let _realCost = system.activePoints / (1 + limitations)